Construct SparseMat using triplets during make_linear_system and fix several bugs of SparseMat and DoubleVec.
- During SparseMat construction, (row#, col#, value) triplets are used
instead of filling SparseMat directly. Since SparseMat has Compressed
Row (or Compressed Column) format which is not efficient for insertion.
Triplets are merged into a SparseMat at the end of construction.

- Implemented a segment_copy method for DoubleVec to copy a part of
another vector instead of using memcpy which is not safe. The new method
calls Eigen's method which should be efficient.