Added a "negate" flag to the Flux class/  This is a hack to make the elastic equation sign change invisible to users.
Flux::negate_ is set by the constructor.  If it's set, the sign of
Flux::output is flipped, and the sign of the normal component is
flipped by NeumannBCApp::integrate.  negate is set only for Stress,
in problem.py.  This compensates for the sign change that was made
to all parts of the elasticity equation, required by the Eigen
solvers which require a positive definite matrix.  Somehow we were
building a negative definite matrix for the force balance equation.
Previous commits made the matrix positive definite, but also flipped
the sign of the output fluxes and stress boundary conditions.  This
flips them back, so that the sign changes are invisible to the user.

These changes haven't been tested yet, because I need to commit them
before I can try to use git to revert the test files.