Changes to some GUI test files required by the elasticity sign changes.
Some tests were failing because the signs of the stress outputs needed
to be changed back to what they were before the elasticity sign change
and the hack to hide it.

Some tests were failing because for some reason they were now reporting
"-0" instead of "0", and gtkMultiTextCompare was using string comparison
instead of floating point comparison.  Adding a tolerance to the
gtkMultiTextCompare call fixes the problem.