Fixed an error in the GUI tests.
Some of the GUI tests were broken when the old CVS headers
were removed, because the line number of the exception raised
in cdebug.C changed, and the line number was part of the error
message that the tests were expecting.  cdebug.C now uses a
hard-coded line number so this problem won't reoccur.