Fixed (I think) the regular expression for setting the version number in make_dist.
The previous version looked for the string
"version_from_make_dist=(unreleased)" in setup.py and oofversion.py,
and replaced "(unreleased)" with the version number.  That doesn't
work, because make_dist changes "(unreleased)" on the master branch,
and so when a development branch is merged, there is no "(unreleased)"
to be changed.

The new version looks for a line beginning with "version_from_make_dist =",
and changes everything after the equals sign to the version number.
So any OOF2 built from a development branch will have
"(unreleased)" for its version number, but an OOF2 built from the
master branch will have a real version number.