Fixed a bug in load_mat() in sparsemat.C
load_mat() was inserting extra 0's into the (0,0) location in sparse
matrices.  Normally, adding 0's to a matrix doesn't have much effect,
but by adding an element, it made the assert(is_lower_trianglular(false))
fail in SparseMat::solve_lower_triangle_unitd(), but only in DEBUG
mode.  That routine isn't used in OOF code anymore, but it's still
tested in the regression suite.  So the test suite caught the least
consequential bug ever.

The extra 0 also had to be removed from the reference file for the
failed test.

Fixed the tmutil line in oof2-clean.