Fixed bugs in Relax.
"Secret" items in LabelTrees aren't displayed in GfxLabelTrees, but
the GfxLabelTree was still trying to undisplay them when they were
removed from the LabelTree.  This caused spurious error messages
in Relax, when the artificial Properties were removed.

The clean up code in Relax.postProcess is now run even if an exception
is raised.

Moved the illegal element check out of CSkeletonElement::cateogoryAreas
and into its caller, CSkeletonElement::c_homogeneity.  Illegal elements
now are assigned category UNKNOWN_CATEGORY, which is -1.