Fixed PyPropertyMethods::py_constant_in_space
py_constant_in_space was trying to call Property::constant_in_space,
which is pure virtual, if the Python Property didn't define it.  The
Python property is required to define it.  g++ 7 was complaining about
undefined symbols when it couldn't find the nonexistant base class
method.

A fix for the SpaceIndex typemap was uncommited on this branch
for some reason.   Maybe I'd copied it from OOF3D.  It just makes
it more robust to exceptions.

Fixed a misspelled comment in pyelasticity.py.