Changed make_dist so that it uses branches instead of tags for the releases.
The old version of make_dist failed because it changed the version
numbers in setup.py and oofversion.py only on the remote master branch,
and subsequent attempts to run make_dist on the next version would
fail because the changes from the remote master hadn't been merged into
the local master.  Now the version numbers will not change on the remote
master, only on the remote release branches.

Reset the version number in oofversion.py to "(unreleased)".  It
had already been reset in setup.py.