Still trying to set the version numbers in the branches.
Something screwy was going on in git.  I probably edited the version
number in setup.py in master, and now can't merge develop into it.
So I merged master into develop, and changed the version number
back to "(unreleased)" in develop, and will then merge that back
into master.  After this, the version nubmer should only be set
in the release branches.