Fixed a problem with images not displaying.
We weren't using ImageMagick routines correctly.  I don't know if
what we were doing used to be correct.  It's possible that ImageMagick
changed and we didn't notice.  This only seemed to affect Mac.

Images aren't being displayed on macOS when using quartz instead
of x11.  This appears to be a bug in gtk2.  Added a note to todo.txt
about upgrading to gtk3.