Merged oof2-build and oof2-clean into one file to avoid code duplication.  Added cocoa info to build and stow directory names.
Although the oof2 cocoa build on the Mac differs from the x11 build
only in the libraries it uses and doesn't use any compilation
parameters, the built files aren't identical, so it makes sense for
them to have different build directories and different stow
directories.  oof2-build now uses 'port installed' to find out
whether we're using x11 or cocoa (quartz) ports, and modifies the
directory names accordingly.

The concatenation of terms into the directory names is now
done more sensibly and won't result in extra hyphens.  This may
change the name of some existing build and stow directories.  You
may want to clean out the old versions.

Because keeping the directory name code in sync between oof2-build
and oof2-clean was getting to be a pain, I merged the two.  oof2-clean
should be deleted, and replaced by a symlink to oof2-build.
oof2-build detects whether it was invoked as oof2-build or oof2-clean,
and does the right thing.