Working on an r3d-inspired version of categoryAreas.
Most of the changes are in pixelsetboundary.[Ch]

This code as committed doesn't work because I forgot that clipping
a loop can create more than one new loop.

The 2D cross product is now a template so it can be used with
Coords, ICoords, or a combination.

A CRectangle can now be constructed from ICoords, not just Coords.

Commented out some old code that doesn't seem to be used:
  class CoordSegment