Starting a branch for burning by EBSD misorientation.
Added a LatticeSystem class (maybe not the best name) that knows
the rotation matrices for crystallographically equivalent orientations.
A singleton is created for each subclass: Cubic, Hexagonal, etc.

Added COrientation::misorientation that computes the misorientation
between two orientations, given a LatticeSystem.

Added a SmallMatrix3x3 subclass of SmallMatrix.  It just adds
easy initialization method, used in the LatticeSystem objects.

Made the Burner base class into a template so that it can be used
on COrientations as well as CColors.