Two changes that make homogeneity calculations significantly faster.
Changed Coord so that it's internally like ICoord (and OOF3D Coord),
using an internal array instead of separate x and y members.
Added an operator[], but didn't yet remove the now deprecated operator().
The thing that was slow was the old implementation of operator(), which
contained a conditional.   All code that explicitly used Coord::x and
Coord::y has been changed to use Coord::x[i].

PxlBdyLoopBase::clip now preallocates space for the new loop data,
guessing that the clipped loop will be the same size as the old
one.  Usually it'll be smaller, so extra memory will have been
allocated, but that's just temporary.  Occasionally it'll be bigger
and the array will have to be reallocated.