Some changes required for gcc 8.2.0 on Ubuntu 18.10.
The arguments to the ltidobject comparator (used to make std::maps
of pointers to instances of IdentifiedObject subclasses) have to be
pointers to the actual subclasses.  They used to be pointers to the
common IdentifiedObject base class.  Now ltidobject has to be a template,
taking the subclass as an argument.