New multi-tiling scheme for pixel set boundaries speeds up the homogeneity calculation significantly.
When constructing the loops around pixels of each category, the
microstructure is divided into tiles, and each tile is done separately.
This allows the element homogeneity calculation to ignore tiles
that don't intersect the element's bounding box, which speeds up
the calculation a lot.  The speedup depends on the size of the
tiles, though.  A skeleton with a wide variety of elmeent sizes has
no single optimal tile size.  Since creating the tiles is cheap,
tiles are now created at many different scales, and the appropriate
scale is used for each element.

Added some commands to the Microstructure menu for setting the
parameters used for choosing the tiling size.  These are only
for debugging and probably won't need to be set by the user.

Added more tests for the homogeneity calculation, to make sure
that it's independent of the tiling size.  Updated some reference
files for the skeleton modification tests, because some are
sensitive to round-off error in the homogeneity values, some of
which have changed by a small amount.

Added operator== for CRectangles.
Added operator<< for printing std::sets.
Deleted some old code.