Added a verbose arg to CSkeletonElement::homogeneiety(), and fixed a bug in the homogeneity calculation.
PSBTiling was not always using the correct tiles when computing
homogeneity.

A comment in skeletoninfoGUI.py marks the line that needs to be
edited to get verbose output from the homogeneity calculation when
an element is clicked in the Skeleton Info toolbox.  The verbosity
is currently limited to the top few function calls, because I didn't
need more to find the current bug.

Most of the changes in this commit are just adding verbose=False
to calls to CSkeletonElement::homogeneity().