Fixes to the misorientation code and Quaterion classes.
There's still somethign wrong with the misorientation calculation.
The lattice symmetry matrices don't make sense.

Rewrote the code for computing quaternions from rotation matrices
by copying code I found on-line. The previous version was badly
behaved if the trace of the rotation matrix was <= -1.

Removed matrices with det=-1 from the symmetry operators.  Some
of the remaining matrices aren't rotations -- I don't understand
the list we were given.

Changed the output operator for SmallMatrix so that it writes
all the values on one line.

Orientation can now compute a misorientation, given a second
Orientation and a lattice system.  There's a "Misorientation
Calculator" menu item in the debug menu that computes misorientations,
mostly for debugging, but it might prove useful for others.