Changed latticesystem.* to use orientation matrices from Mathematica.
Matrices are stored in a LatticeSymmetry object which is stored
in a std::map keyed by the Schoenflies symbol.

This version is crashing on startup in the next python import
after latticesystem.py, so there must be a stray pointer somewhere.
I'm committing now so I can transfer to Linux and run valgrind.