Lattice symmetry can be specified by Space Group number, Schoenflies symbol, or in international notation.
There are Enum classes for point groups in the Schoenflies and
International notations.  Space groups are identified by integer
numbers 1-230.  Convertible registered classes convert between the
three, always choosing the lowest numbered space group when converting
from a point group.

The C++ function getLatticeSymmetry takes a std::string arg which
is a Schoenflies symbol, and returns a LatticeSymmetry object, which
contains a vector of rotation matrices.