Fixed bugs in Rationalize.
The RemoveBadTriangle Rationalizer was not setting parent/child
relationships for new segments when applying its triTriSplit
method on the boundary of a Skeleton with periodic boundary conditions.

Changed the way RemoveBadTriangle chooses the specific rationalizations
to apply.  It used to try all of them (splitting obtuse angles, collapsing
acute angles) if the element failed either the obtuse or acute angle tests.
Now it only tries the fixes relevant to the failed tests.

Changed the way ProvisionalChanges handles segment substitution
so that it doesn't create new SkeletonSegments until a change
is accepted.

Added progress bars for the long-form Skeleton sanity test.