Further fixes to skeleton rationalization.
The fix that was required when splitting a triangle abutting a
triangle across a periodic boundary was also required for triangles
abutting quadrilaterals across a periodic boundary.

Restored the old behavior, in which the elements created by
rationalizing are also rationalized in the same pass.

Updated a test reference file which changed in nonessential ways
with the new rationalization algorithm.