Debugging AutoGrain, which runs Burn on randomly selected pixels until all pixels are in a group.
Currently it segfaults, so I'm commiting here to transfer to Linux and valgrind.

The autograin code is currently all in image/autograin.*, which is
the wrong place for it, since it will eventually work on EBSD data
too.

Copied the new random.[Ch] from OOF3D, because it has a working shuffle().

Made SimpleArray2D in array.h a little less simple.  It can now use
an ICoord as a constructor or index arg, and has a conatains() method.