New "Auto" button on the Microstructure page. Removed AutoGrain from Image page.
The "Auto" button on the Microstructure page can create either contiguous
or discontiguous groups based on color similarity or orientation similarity.
It technically makes the AutoGroup button on the Image page obsolete,
but we might want to keep it anyway.

The new methods appear to work, but there are some GUI glitches to fix.
The contiguous and discontiguous selectors need different PixelDifferentiator
classes, because when making discontiguous selections there's no concept
of local vs global flammability, as in the Burn algorithm.  The RCF for
the PixelDifferentiators currently shows all of them, but it should show
only the ones appropriate for the selected contiguity.