Working on a new AutoGroup command.
The new method should work to group similar but not exactly equal
pixel values, making it more useful when grouping by orientations
or on images that weren't artificially generated.  The current
implementation seems to work basically, but will need some testing
and tweaks.  The main routine is statgroups() (because it's sort
of statistical) in statgroups.C.  It's a stupid name but I haven't
thought of a better one.

The button for the new routine is the "Auto2" button on the Microstructure
page, for now.  It'll get a better name or be merged into the old Auto
button or something, later.

Grouping by orientations isn't working yet, but I added the necessary
arithmetic to the COrientation class.  The pixel values to be grouped
have to be able to compute a weighted average.  It doesn't make sense
to add orientations, but it does make sense to average them, I think.