Copied fixes from color autogrouping to orientation autogrouping.
If two orientations are identical, their misorientation is now
exactly 0, instead of numerical error.  This makes it possible
to autogroup piecewise constant EBSD data correctly.

When adding a pixel to a group, the mean doesn't change if the pixel's
orientation is exactly at the mean.

If the variance is less than var0, it's reset to var0, so that
it won't become 0 when a second pixel is added to a group.

Commented out some debugging lines, but this time I didn't remove
them, because I keep adding them back in.