Added a PropertyOutput for Orientation.
Orientation output currently only works for the Orientation Property,
not for orientation maps.  That shoudl be easy to add at this point.

COrientation is now an OutputVal, so that it can be used by the
output machinery.  It's a different kind of OutputVal, though.  It
doesn't make much sense to use it in anything other than direct output,
because most of the arithmetic used by other output methods can't be
easily done on orientations.  So COrientation is a subclass of
NonArithmeticOutputVal, for which the arithmetic operations defined
for other OutputVals just throw an exception.  (The right way to do
this will be to also define an ArithmeticOutputVal, and only define
the arithmetic operations there.  Also, the GUI should prevent
NonArithemeticOutputVals from being used in output methods that
don't make sense for them.)

Moved corientation* from SRC/common to SRC/engine, because outputval.*
is in engine.

Cleaned up some DIR.py files by removing the irrelevant 3D lists
and sorting the 2D ones alphabetically.  I was bored.

Fixed a bug in MetaRegisteredParameter, which was missing its
clone() method.

Removed some commented-out code from direction.C.  That code was
the only reason that corientation.* had been in common instead of
in engine.

Added methods to RegisteredClass for fetching Registrations
when don't actually have an instance to work with.

The Orientation classes in orientationmatrix.py now generate
Enum entries when they're registered.  The Enum is used to
generate the list of available orientation formats that can
be chosen when generating an orientation output.

Removed the vestigial elasticity.spy.  It imported some modules
and didn't do anything with them.