Added an AngleRangeParameter and widget.
AngleRangeParameter differs from FloatRangeParameter in the way it
handles values that are outside of its range.  If the range covers
a full circle, AngleRangeParameter adds or subtracts 2pi (or 360)
to get the value back where it wants it.

LabelledSlider can now be told how to handle values that are out
of its range, via a new optional clipperclass constructor argument.