Improvements to the SimpleArray2D template.
It no longer initializes itself with memset.  The clang compiler
warned that memset would overwrite the vtable when the template
argument is a dynamic type.

Data is now stored as a std::vector.

Data is retrieved using std::vector<TYPE>::reference, instead of
simply TYPE&, because TYPE& doesn't work if TYPE==bool.  std::vector<bool>
is a specialization that stores bits, not bytes.