Still working on output for property parameters.
Committing now to have a reference point to go back to after
eliminating PropertyOutputInit::operator(), which creates an OutputVal
inside PropertyOutput::evaluate.  The ListOutputInit doesn't have
enough information at that point to create the right size ListOutputVal.
But it should be possible to create the OutputVal before calling
evaluate, and just clone it when needed.