Working being able to output moduluses and other Property parameters.
Reworked the way OutputVals are initialized.

The class structure is a little bit less confusing now.  It still
needs better documentation.

There's now a default initializer for OutputVals, and the specialized
ScalarPropertyOutputInit, TwoVectorPropertyOutputInit, etc, aren't
used any more.  The default initializer assumes that the
PropertyOutputRegistration has a zeroVal method which returns an
OutputVal representing zero.  The initializer then clones that
object at each evaluation point.  It's still possible to provide a
nongeneric initializer, such as for Orientations, which can use the
Output's parameters to figure out what object to return.

The current version passes the output_test.py test.  It can output
orientations. It still doesn't properly output Cijkl, but that's
not in the tests yet.