Fixed an apparently unused method and deleted the comment in a related method asking why it didn't do the same thing as the unused method.
double &SymmMatrix3::operator[](const IndexP&) does the conversion
from IndexP to SymTensorIndex in two steps.  The const version of
the same method did it in one.  After changing the non-const version
so that it converted in one step, the program compiled but crashed
at run time on macOS, but not Linux.

Removing the const version of the routine (and the declaration in
the abstract base class) doesn't affect the program, so it's
apparently not used at the moment.  I changed it to use the
two-step index conversion anyway.