Output improvements.   Outputs can be concatenated, and Material Constants can be displayed in the Mesh Data viewer.
OutputParameters have incomputable() methods so the GUI knows when
they can be displayed.

OutputVals need to be iterated over and need to be able to return
a particular component when asked. For example, the GUI needs to
display the "alpha", "beta" and "gamma" components of a COrientABG,
but differently named components of a COrientQuaternion.  History
requires the index and iteraor to be subclasses of FieldIndex and
FieldIterator.  corientation.h defines templates for indices and
iterators for the COrientation subclasses.

Using FieldIndex for the above is ugly because it makes some
field-specific assumptions (namely, that planarity matters).  It's
name should be changed, and a new FieldIndex derived from it, but
that's a low priority.

FieldIterators now know their sizes.

ListOutputVals are now initialized with a list of labels, one for
each component, not just a size, and can be indexed and iterated.