PropertyOutputs throw an exception if the material isn't consistently defined.
Without this, it was possible to try to compute an Output from a
Material with an anisotropic property but no orientation property.
This would crash if the property had to be converted into the lab
frame.

It's sometimes possible to compute a meaningful Output from a badly
defined Material, but detecting those cases is difficult.  It's
easier and not unreasonable to insist that the Material be well
defined.