Fixed some output bugs.
Clicks on the "Aggregate" button on the Analysis Page hadn't been
connected to a callback, which meant that the weren't logged for
GUI tests.  The callback was still being called, though, because
it was called when the "Scalar" button was unclicked.  This was
putting a checkpoint in the log file that wasn't being reached on
replay.  Now both buttons are logged, but the callback checks
to make sure it's not running twice.

All DataOperationRegistrations now have an outputFilter that
decides if they can be applied to an Output.