Partway through updating the image directories for gtk3.
This currently doesn't work, partly because the non-gui part of
image now requires the gui part of common, because OOFCanvas
isn't properly separated between gui and non-gui pieces.  As
a result, the non-gui part of image won't build because it requires
gtk and setup.py isn't telling it how to find it.

I have refactored OOFCanavs so that its gui parts can be separated,
but am having trouble building it in its test directory.  It should
be easier to build in OOF2, using the setup tools we've developed, so
I will finish converting SRC/image to Gtk3 and debug the OOFCanvas
refactoring at the same time, and worry about making a stand-alone
refactored OOFCanvas later.