The initial contents of the graphics window are now zoomed correctly.
There was a race condition between zooming and drawing the initial contents
of the window, due to the zooming being triggered by the gtk "draw"
signal and the drawing being done by the buffered output device.  Since
buffering is now now necessary and the output device classes will be
going away, the canvas now uses the CanvasOutput device direclty,
without the BufferedOutputDevice.