The image is flipped before being sent to the Canvas.
We flip the image when loading it so that we can use a right handed
coordinate system, but that means that it has to be flipped back
for display.  In the old Gtk2 version we used our own code when
converting the ImageMagick image to a displayable form for
libgnomecanvas, and flipped it back in the process.  The new OOFCanvas
code extracts the image data directly from ImageMagick, so we need
to flip it explicitly.