Moved Andrew's addition to the 2.1.19 section, because I wasn't clear about the protocol that I've been using.
Maybe it makes more sense to put the latest modifications above the
version number, since we don't know what the version number will
be when the modifications are eventually released, but I'm not going
to overthink this right now.  I just didn't want the record of the
two mods already made to be lost.