Mostly minor rubberband related fixes.
GfxWindow.setRubberband checks that the canvas exists, to avoid
crashing during window destruction.

CanvasRectangle draws its line on the inside edge of the rectangle,
to be more like CanvasCircle.  This also means that he bounding box
doesn't change when the line width changes (unless the line width
is greater than the rectangle size, which is dumb.)

Added a MicrostructurePerimeter display, which just draws a
rectangle around the Microstructure.  I needed something simple
to display while debugging the rubberband issues.