Added a non-antialiased way to display images, and made it the default.
There's currently no way to switch back to the other method, but that
will come later.

The new way doesn't simply copy the CanvasImage's Cairo::ImageSurface
to the display, but it draws individual pixels as filled rectangles.  For
the simple cases I've tested so far, it's not terribly slow, but it does
some very strange things when composited with another image, as when marking
selected pixels.  It does solve the fuzziness problem, but at some zoom levels
the aliasing is bad.