Comments about some graphics layer issues to be fixed later.
Editing graphics layers isn't working properly, and CanvasLayers aren't
being deleted reliably.  The issues are related, and both will be
simpler when the LayerEditor and LayerSets are removed, as in OOF3D.
So I'm going to ignore those problems for now, finish the conversion
of all GUI code to Gtk3, and then get rid of the LayerEditor and
LayerSets, all before investigating the editing and deletion issues.