More cosmetic improvements and a few substantial ones.
Column synchronization in the Scheduled Output page is now better
than it was with gtk2, but still not perfect.  It's not obvious
to the user how to scroll the views, since the scrollbar only
appears in the rightmost pane.  I'll try making it more prominent
next.  At least it doesn't completely mess up if the font size
is changed anymore.