Working on a new version of the Scheduled Output page.
I started reworking the old version of the page to resolve some
scrolling issues, but decided that the best way to make sure that
the four columns scrolled together was to put them all in a
single TreeView.  They had been in separate TreeViews so that each
could have its own set of buttons.  But that made it hard to keep
the rows synchronized when scrolling, especially if the user changed
the font size (since one of the columns contained no text and didn't
change its row spacing).

The new version of the page is now installed, and the old one isn't,
but I haven't gone so far as to delete its code yet.  The new version
still needs some rearranging.