Bug fixes and enhancements to skeleton and mesh display methods.
The default line width for mesh and skeleton elements is now non-zero.
Apparently zero-width lines in Cairo aren't visible.  They were visible
in libgnomecanvas.

Selected nodes and segments are now marked in two colors -- a dot or
line of the desired size and color on top of a slightly wider white
dot or line.  This makes them visible against more backgrounds.
Perhaps this should be optional.  I didn't make the same change for
the full skeleton and mesh displays because it might be too busy.