Various mostly minor fixes.  Addressed some TODOs.  Fixed some more widget spacing issues.
FramedChooserListWidget and ScrolledChooserListWidget now pass
kwargs through to their Gtk.Frame.

The list of groups in the Microstructure page doesn't try to cram
itself into the top two lines of the space it has available.

Fixed sorting in the Boundary Condition page.  Removed the commented-out
code for sorting by the enabled status, but left in the comment about
why it's a bad idea.