Addressed a few more TODOs.
The initial size of the console window is now set, but not very
elegantly. It's just set to the initial width of the OOF2 window.
I don't know how to set it in terms of the character size, the way
it's done in gtk2 and the way Gtk.Entrys can be handled in gtk3.
I left in a TODO, but downgraded it.

Added a utility function to runtimeflags to make it easier to get
the initial size of the OOF2 window.

Added to the TODO in oof_mainiteration.py but didn't actually do
anything about it.  I won't know if it's still necessary until I
work on animation in the graphics window.

Changed the way that empty ParameterTables work, and adjusted
spacings in RegisteredClassFactory.  RCF's with no parameters looked
funny because they were leaving space between the ChooserWidget and
the empty ParameterTable, with the effect of leaving too much blank
space below the Chooser.

Deleted some TODOs that don't have to be done after all.