Fixed a bug in the RegisteredClassFactory.
If a widget that used keyboard input was used inside a
RegisteredClassFactory, and the widget was part of a signalling
loop that caused the factory to be updated (eg, the Entry in a
LabelledSlider that synchronizes its slider and entry) then the
factory would rebuild the widgets on each pass through the loop,
causing the keyboard focus to be lost.  The solution is to check
to see if the registration has changed and not to rebuild the widgets
if not necessary.