Fixed a ChooserWidget misfeature, and a bug in the PixelAggregateWidget.
In the gtk2 version, if ChooserWidget.set_value() was called with an
argument that wasn't either None, an integer in the appropriate range,
or a string in the widget's list of choices, it failed silently,
sometimes (but not always) selecting the first item in its list.  Now
it raises an ErrPyProgrammingError.

The aforementioned ChooserWidget feature was hiding a bug in the
PixelAggregateWidget, which is derived from ChooserWidget. If it
tried to initialize itself to placeholder.selection or placeholder.every,
it would fail silently.  Now it passes the placeholder's IDstring
to set_value() instead of the placeholder itself.