Removed the Layer Editor window.
The GUI is now like OOF3D:  the Layer/New menu item in a graphics
window opens a dialog that lets you create a new layer, which is
sent to the same graphics window.   Double clicking a layer
or using Layer/Edit opens the same dialog box, but its parameters
are initialized with the selected layer, and the new layer replaces
the old layer.

Along with the layer editor, the Display and LayerSet classes are
now defunct.

There are still a lot of glitches in the graphics window but it's
mostly working.