Fixed problems with zooming to fill the window.  A margin can be set in the OOFCanvas.
Changed the way bounding boxes are handled in CanvasItems.  Now a
bbox must be passed to the base class constructor.

Any item can now have components with sizes specified in pixels.  Got
rid of the PixelSized subclass for CanvasItem.

CanvasItems whose size in physical units depends on the pixel size
must redefine pixelExtents(), which computes how far in pixels
the item extends past its bounding box.

Setting the margin on a canvas now works.

When zooming out, if the image is smaller than the widget it is
centered, rather than being jammed into the top left corner.   This
is done by making the underlying bitmap always at least as large
as the widget.

There are still some peculiarities.  If you zoom out and back in,
the scroll region is too large, and you can scroll the visible
objects off the screen entirely.