Fixed some misleading names.  Trying to fix rubberband drawing.
Ever since adding the canvas margins and centering the image when zoomed out, the rubberbands haven't been drawn correctly when the scroll bar adjustments are nonzero.
The current implementation is straightforward and really ought to work but there's still an incorrect offset in the x direction only.  That makes no sense.
I'm committing now so that I can see if it does the same nonsense on Linux.