Fixed a test problem revealed by the new zstrain_test.
zstrain_test wasn't working when run as part of the full test
suite.  That was because an earlier test hadn't deleted its
Microstructure, so when zstrain_test created a new Microstructure
it was called "microstructure<2>", and not actually used in the
test.

zstrain_test now uses the memorycheck.check decorator, which deletes
the Microstructure and checks for memory leaks.  zstrain_test no
longer is responsible for deleting its boundary conditions because
memorycheck.check does it.