Fixed some problems building swig, sort of.
"python setup.py clean" wasn't deleting all of the intermediate
files created when building swig.  If the version of bison or yacc
changed between oof2 builds, then swig could be built incorrectly.
Now "python setup.py clean" runs "make clean" in the swig source
directory, and "make clean" deletes all of the generated files.

It still would be a good idea for the intermediate files created
when building swig to live in the OOF2 build directory instead
of the OOFSWIG directory, but this makes the consequences of not
doing that easier to fix.