Fixed a couple of pixel selection bugs.
The Brush selector was being too aggressive in removing possibly
redundant points from the list of points to be processed.  Now it
just uses all the points.

The GenericSelectToolboxGUI was sometimes not including the mouse
up point in the list of points to be processed.  This may have been
due to a difference between gtk+2 and gtk3.

Also fixed a recently created bug which tried to set rubber band
parameters without checking that a rubber band existed.