Added scrollwheel support in OOFCanvas, and basic use of it in GfxWindow.
Using the sroll wheel (two-finger swipe) in the graphics window
scrolls the image as it should.  What it doesn't do is log it, so
scripts that rely on scrollwheel events won't work.  Logging is
difficult because we don't want to log every scroll event, and there's
no event that's emitted indicating the end of the scrolling process.