Simplified some graphics layer handling code, added some TODOs, and fixed some bugs.
There had been a lot of code duplication in the bookkeeping calls
that various layer manipulations routines did after they changed
the layers.  Now they call call GhostGfxWindow.layersHaveChanged(),
which is slightly redefined in GfxWindowBase.  This simplifies the
code and fixes some bugs.

The code for sorting layers and deciding when they needed to be
sorted wasn't working.  Now it's better.  There's no flag that
has to be set when the layers are unsorted.  The routine that
sensitizes the "Reorder All" menu item just checks the layers
themselves to see if they're in order.

Some of the Skeleton and Mesh display methods were assuming that
coordinates were Python Points when sometimes they can be swigged
C++ Coords.  I'm not sure at the moment why the problem cropped up
now, but all display methods in engine/IO/displaymethods.py now use
pt[0] and pt[1] instead of pt.x and pt.y, which works for both
Points and Coords.