Removed the menu routines for manipulating the scroll bars in the canvas.
It's difficult to create menu routines for logging and reproducing
scroll wheel events, and it's inconsistent to log the scroll bars
but not the scroll wheel.  There are no menu items that rely on the
actual position of the canvas, so it makes sense to just get rid
of the scrolling menu.   (Scroll wheel events can't be logged because
there's nothing that signals the end of a series of events, and the
individual events are too numerous to log.)