Saving the contour map to pdf works, even in batch mode.
Contour map data and some functionality was moved from GfxWindow
to GhostGfxWindow so that the contour map canvas can be created and
saved when the GUI isn't active.

The OOFCanvas::Canvas::saveAsPDF methods return a bool indicating
success, rather than crashing or throwing an exception.  They had
been crashing if the canvas was empty.