Fixed one rubber band bug, and apparently added a new one.
The rubber band is now drawn on a mouse down event, without waiting
for the first mouse-move event.  This is important for the brush
selectors, so that the user can see the size of the brush before
moving it.

For some reason, that change is making the Spider rubberband, used
when moving Skeleton nodes, to seg fault.  Committing this now so
that I can try valgrind on linux.