Fixed problems with the rubber band for the Move Node toolbox, and added a TODO.
The endpoints for the rubberband had to be determined by querying
the Skeleton, which was done on a subthread so that the Skeleton
could be locked.  That meant that the mouse down callback was
returning before the rubberband was created.  Now the rubberband
is created before starting the subthread, and the endpoints are
added later.

Added a TODO about preventing this possibility by requiring that
the mouse down callback return a rubber band if it wants to use
one.