Better control of when mouse motion events are generated by the canvas, and a toolbox bug fix.
Instead of just having a bool indicating whether or not motion
events should call the callback, the Canvas now uses an enum with
three values: MOTION_NEVER, MOTION_ALWAYS, and MOTION_MOUSEDOWN.

The mode selection buttons on the Skeleton selection toolbox were
calling their callback twice, both when a button was activated
and when it was deactivated.  This was confusing when trying to figure
out whether the changes to mouse motion handling were working.